+++ apps/cow-fi/hooks/useSetupPage.ts
@@ -1,6 +1,6 @@
 'use client'
 
-import { useInitializeUtm } from "modules/utm"
+import { useInitializeUtm } from 'modules/utm'
 
 export function useSetupPage() {
   useInitializeUtm()
